+++ src/main/java/roomescape/dao/ThemeDao.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.dao;
 
 import java.time.LocalDate;
