+++ src/main/java/roomescape/web/dto/request/ReservationRequest.java
@@ -1,13 +1,14 @@
-package roomescape.dto.request;
+package roomescape.web.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+import roomescape.service.command.ReservationCommand;
 
 import java.time.LocalDate;
 
-public record WaitingRequest(
+public record ReservationRequest(
         @NotBlank(message = "이름은 필수입니다.")
         String name,
 
@@ -21,4 +22,12 @@ public record WaitingRequest(
         @Min(value = 1, message = "유효한 테마 ID가 필요합니다.")
         long themeId
 ) {
+        public static ReservationCommand toReservationCommand(ReservationRequest request) {
+                return new ReservationCommand(
+                        request.name,
+                        request.date,
+                        request.timeId,
+                        request.themeId
+                );
+        }
 }
