+++ src/main/java/roomescape/web/dto/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto.response;
+package roomescape.web.dto.response;
 
 public record ErrorResponse(String message) {
 }
